mlops-multi-account-cdk/mlops-sm-project-template/seed_code/iot_deploy_app/app/requirements.txt
absl-py==1.4.0
astor==0.8.1
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
Flask==2.2.3
flatbuffers==23.1.21
gast==0.5.3
h5py==3.8.0
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.2.1
numpy==1.24.2
onnx==1.13.1
onnxruntime==1.13.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
requests==2.28.2
six==1.16.0
sympy==1.11.1
termcolor==2.2.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
